ai-service/requirements.txt
# WarrantyVault AI Microservice Dependencies

# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OCR and Document Processing
paddleocr==2.7.0.3
paddlepaddle==2.6.0
pdf2image==1.17.0
Pillow==10.2.0

# AI/ML for structured extraction
transformers==4.37.2
torch==2.1.2
sentencepiece==0.1.99

# Utilities
python-dotenv==1.0.1
